Add gradle, refactor other manifest commands
